Refactor the code for creating tag

+++ src/main.rs
@@ -490,9 +490,24 @@ fn diff() -> io::Result<()> {
 fn tag(tag_name: &str) -> io::Result<()> {
     // TODO: Change heard code head_file value (currently master) to dynamic according to active branch
     let head_content = fs::read_to_string(".rgit/refs/heads/master")?;
-    let tag_file_path = format!(".rgit/refs/tags/{}", tag_name);
+    let tag_folder_path = Path::new(".rgit/refs/tags");
+
+    // If tags folder not created yet, then create it
+    if !tag_folder_path.exists() {
+        fs::create_dir_all(tag_folder_path)?;
+    }
+
+    // Create new tag file inside tags dir
+    let tag_file_path = tag_folder_path.join(tag_name);
+    // If tag file name already exist them skip the creation
+    if tag_file_path.exists() {
+        println!("Tag '{tag_name}' already exist");
+        return Ok(());
+    }
 
     let mut tag_file = File::create(tag_file_path)?;
+
+    // Store latest hashed commit blob to new tag file
     tag_file.write_all(head_content.as_bytes())?;
 
     println!("Tag '{}' created", tag_name);
